refactor: update the default sign_data
